"add wrong path and some list data"

+++ frontend/src/components/Admin/ControlUser.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import { Button, Input, Layout, notification, Space, Table, Typography } from 'antd';
+import { Input, Layout, Popconfirm, Table, Typography, message } from 'antd';
 import { UserInterface } from '../../models/User';
 import { DeleteUser, GetUserList } from '../../services/HttpServices';
 import { ColumnsType } from 'antd/es/table';
@@ -7,17 +7,29 @@ import { DeleteOutlined, SearchOutlined } from '@ant-design/icons';
 import moment from 'moment';
 
 const ControlUser: React.FC = () => {
+  let admin = localStorage.getItem("name")
+  const { Content } = Layout;
+  const { Title } = Typography;
   const [user, setUser] = useState<UserInterface[]>([]);
   const [searchText, setSearchText] = useState<string>('');
-  const [api, contextHolder] = notification.useNotification();
+  const [confirmation, setConfirmation] = useState('');
+  const [api, Holder] = message.useMessage();
   const filteredUser = user.filter((userData) =>
     userData.Status !== 0 &&
     (
+      userData.Firstname.toLowerCase().includes(searchText.toLowerCase()) ||
+      userData.Lastname.toLowerCase().includes(searchText.toLowerCase()) ||
       userData.Nickname.toLowerCase().includes(searchText.toLowerCase())
     )
   );
-  const { Content } = Layout;
-  const { Title } = Typography;
+
+  const openAlert = (type: 'success' | 'error', content: string) => {
+    api.open({
+      type,
+      content,
+      duration: 5,
+    });
+  };
 
   const getuserlist = async () => {
     let res = await GetUserList();
@@ -26,37 +38,13 @@ const ControlUser: React.FC = () => {
     }
   };
 
-  const checkDelete = async (id?: number) => {
-    const key = `open${Date.now()}`;
-    const close = () => {
-      console.log('Notification closed');
-    };
-    const btn = (
-      <Space>
-        <Button type="link" size="small" onClick={() => api.destroy()}>
-          Close
-        </Button>
-        <Button
-          type="primary"
-          size="small"
-          onClick={() => {
-            api.destroy(key);
-            DeleteUser(id);
-            window.location.reload();
-          }}
-        >
-          Confirm
-        </Button>
-      </Space>
-    );
-    api.open({
-      message: `ลบข้อมูล`,
-      description:
-        `คุณต้องการจะลบข้อมูลใช่หรือไม่ ?`,
-      btn,
-      key,
-      onClose: close,
-    });
+  const handleDelete = async (id?: number) => {
+    if (confirmation === admin) {
+      await DeleteUser(id);
+      getuserlist();
+    } else {
+      openAlert('error', 'ชื่อไม่ตรง ไม่สามารถลบได้');
+    }
   };
 
   const columns: ColumnsType<UserInterface> = [
@@ -117,44 +105,57 @@ const ControlUser: React.FC = () => {
       width: '10%',
       align: 'center',
       render: (text, record) => (
-        <span className="icon-table">
-          <DeleteOutlined
-            style={{
-              fontSize: '20px',
-              justifyContent: 'center',
-              alignItems: 'center',
-              color: '#1890ff',
-            }}
-            onClick={() => checkDelete(record.ID)}
-          />
-        </span>
+        <>
+          {Holder}
+          <Popconfirm
+            title={
+              <>
+                <Input
+                  placeholder="พิมพ์ ชื่อจริง เพื่อยืนยัน"
+                  onChange={(e) => setConfirmation(e.target.value)}
+                />
+              </>
+            }
+            onConfirm={() => handleDelete(record.ID)}
+            cancelText="ไม่"
+            okText="ใช่"
+            icon={null}
+          >
+            <DeleteOutlined
+              style={{
+                fontSize: '20px',
+                justifyContent: 'center',
+                alignItems: 'center',
+                color: '#ff4d4f',
+                display: 'flex',
+              }}
+            />
+          </Popconfirm>
+        </>
       ),
     },
   ];
-  
+
   useEffect(() => {
     getuserlist();
   }, []);
 
   return (
-    <>
-      {contextHolder}
-      <Layout style={{ minHeight: '100vh' }}>
-        <Content style={{ margin: '16px' }}>
-          <div style={{ padding: 12, background: '#fff', minHeight: 360, textAlign: 'center' }}>
-            <Title level={3}>ตารางรายชื่อผู้ใช้งานระบบ</Title>
-            <Input
-              prefix={<SearchOutlined />}
-              placeholder="ใส่ชื่อเล่นเพื่อค้นหา"
-              value={searchText}
-              onChange={(e) => setSearchText(e.target.value)}
-              style={{ marginBottom: 16 }}
-            />
-            <Table columns={columns} dataSource={filteredUser} />
-          </div>
-        </Content>
-      </Layout>
-    </>
+    <Layout style={{ minHeight: '100vh' }}>
+      <Content style={{ margin: '16px' }}>
+        <div style={{ padding: 12, background: '#fff', minHeight: 360, textAlign: 'center' }}>
+          <Title level={3}>ตารางรายชื่อผู้ใช้งานระบบ</Title>
+          <Input
+            prefix={<SearchOutlined />}
+            placeholder="ใส่ชื่อจริง ชื่อเล่นหรือนามสกุล เพื่อค้นหา"
+            value={searchText}
+            onChange={(e) => setSearchText(e.target.value)}
+            style={{ marginBottom: 16 }}
+          />
+          <Table columns={columns} dataSource={filteredUser} />
+        </div>
+      </Content>
+    </Layout>
   );
 };
